SA1649: The file name of a C# code file does not match the first type declared in the file.
